Automatically builds app code deb package for staging
The staging hosts should install the SecureDrop app code from the code
in the local git repository, rather than pulling down debs from the FPF
apt repository. The build-deb-pkgs.yml playbook is now included at the
beginning of the staging playbook.

The build-deb-pkgs.yml playbook has some clunky var_files includes,
because the Ansible group management in the Vagrantfile is difficult
to override cleanly. Not willing to drop Vagrant support for these
hosts, so we'll just have to live with the hardcoded includes for now.